Fixed up ArticleInfo to match Kelsey's new category format.
Synched Kelsey's category files between the client and server package.
Did my best to eliminate compile errors that have been cropping up with everyone's commits but there are still a lot on the client.  Server should be mostly free of them as long as that database library reference is resolved and the changes to ArticleInfo are accounted for.